Fix credential leakage in logging
The subprocess exception traceback prints the called command, and this
includes the database password in the case of the pg_dump call.

We switch to Logger.error() from exception() to fix this leakage. The
traceback from this particular point is not useful anyway, and even with
error(), the called command still emits the more useful error messages.

This is technically a security issue but it should have no known impact
as this code's repository was published only less than an hour ago and
the code is not known to be used in any production system.